backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
python-dotenv==1.0.1
httpx==0.27.0
requests==2.31.0
numpy==1.26.4
scikit-learn==1.4.2
lightgbm==4.3.0
sentence-transformers==2.7.0
transformers==4.40.1
torch==2.3.0
umap-learn==0.5.6
hdbscan==0.8.38.post1
supabase==2.4.3
python-multipart==0.0.9
